Updated WPILib source code for 2010
git-svn-id: https://kwarqs-dev.bbn.com/svn/kwarqs/code@375 eb211378-2a44-443b-ae4a-d45323d45073

+++ WPILib/DigitalModule.h
@@ -39,7 +39,7 @@ public:
 	UINT16 GetDIO(void);
 	bool GetDIODirection(UINT32 channel);
 	UINT16 GetDIODirection(void);
-	void Pulse(UINT32 channel, UINT8 pulseLength);
+	void Pulse(UINT32 channel, float pulseLength);
 	bool IsPulsing(UINT32 channel);
 	bool IsPulsing();
